moved audio components to separate folder
