+++ src/components/audio/AudioEqualizerStyles.ts
