+++ src/components/audio/AudioPlaybackStyles.ts
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  playbackProgressBarContainer: {
+    alignSelf: "stretch",
+    width: "100%",
+  },
+  playbackButtonsContainer: {
+    alignItems: "center",
+    columnGap: 8,
+    justifyContent: "center",
+  },
+  playbackProgressBar: {
+    alignSelf: "stretch",
+    height: 14,
+    marginHorizontal: 0,
+    marginVertical: 0,
+  },
+});
