+++ src/components/audio/AudioUtils.ts
@@ -1,9 +1,4 @@
-export const formatRecordingDuration = (
-  durationMillis: number,
-): string | null => {
-  if (!Number.isFinite(durationMillis) || durationMillis < 0) {
-    return null;
-  }
+export const formatRecordingDuration = (durationMillis: number): string => {
   const totalSeconds = Math.max(0, Math.floor(durationMillis / 1000));
   const seconds = totalSeconds % 60;
   const minutesTotal = Math.floor(totalSeconds / 60);
@@ -13,9 +8,12 @@ export const formatRecordingDuration = (
   const secondsStr = String(seconds).padStart(2, "0");
   const minutesStr = String(minutes).padStart(2, "0");
 
+  const parts = [];
   if (hours > 0) {
-    return `${hours}:${minutesStr}:${secondsStr}`;
+    parts.push(String(hours));
   }
+  parts.push(minutesStr);
+  parts.push(secondsStr);
 
-  return `${minutesStr}:${secondsStr}`;
+  return parts.join(":");
 };
