+++ src/components/audio/index.ts
@@ -1,0 +1,3 @@
+export { AudioEqualizer } from "./AudioEqualizer";
+export { AudioPlayback } from "./AudioPlayback";
+export * as AudioUtils from "./AudioUtils";
